Updated AWS API's to V2, other various bugfixes and improvements.

+++ internal/gosbench/driver/workItems.go
@@ -6,16 +6,15 @@ import (
 	"fmt"
 	"math/rand"
 	"sort"
+	"sync"
 	"time"
 
+	"github.com/aws/aws-sdk-go-v2/feature/s3/manager"
 	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/aws/request"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/s3/s3manager"
-	log "github.com/sirupsen/logrus"
+	"github.com/rs/zerolog/log"
 )
-
-var RandomData []byte
 
 // WorkItem is an interface for general work operations
 // They can be read,write,list,delete or a stopper
@@ -50,13 +49,15 @@ type WriteOperation struct {
 
 // ListOperation stands for a list operation
 type ListOperation struct {
-	TestName       string
-	Bucket         string
-	ObjectName     string
-	ObjectSize     uint64
-	MPUEnabled     bool
-	PartSize       uint64
-	MPUConcurrency int
+	TestName                 string
+	Bucket                   string
+	ObjectName               string
+	ObjectSize               uint64
+	WorksOnPreexistingObject bool
+	MPUEnabled               bool
+	PartSize                 uint64
+	MPUConcurrency           int
+	UseV2                    bool
 }
 
 // DeleteOperation stands for a delete operation
@@ -98,10 +99,6 @@ func GetNextOperation(Queue *Workqueue) string {
 	return Queue.OperationValues[0].Key
 }
 
-func init() {
-	WorkContext = context.Background()
-}
-
 var WorkContext context.Context
 
 // WorkCancel is the function to stop the execution of jobs
@@ -119,85 +116,104 @@ func IncreaseOperationValue(operation string, value float64, Queue *Workqueue) e
 }
 
 // Prepare prepares the execution of the ReadOperation
-func (op ReadOperation) Prepare() error {
-	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).WithField("Preexisting?", op.WorksOnPreexistingObject).Trace("Preparing ReadOperation")
+func (op *ReadOperation) Prepare() error {
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Bool("Preexisting?", op.WorksOnPreexistingObject).Msg("Preparing ReadOperation")
 	if op.WorksOnPreexistingObject {
 		return nil
 	}
 	if op.MPUEnabled {
 		if op.PartSize == 0 {
-			op.PartSize = s3manager.DefaultDownloadPartSize
+			op.PartSize = manager.DefaultDownloadPartSize
 		}
 		if op.MPUConcurrency == 0 {
-			op.MPUConcurrency = s3manager.DefaultDownloadConcurrency
+			op.MPUConcurrency = manager.DefaultDownloadConcurrency
 		}
-		return putObjectMPU(HousekeepingSvc, op.ObjectName, bytes.NewReader(RandomData[:op.ObjectSize]), op.Bucket, op.PartSize, op.MPUConcurrency)
+		uploader := manager.NewUploader(HousekeepingSvc, func(d *manager.Uploader) {
+			d.PartSize = int64(op.PartSize)
+			d.Concurrency = op.MPUConcurrency
+		})
+
+		return putObjectMPU(uploader, op.ObjectName, bytes.NewReader(generateRandomBytes(op.ObjectSize)), op.Bucket)
 	} else {
-		return putObject(HousekeepingSvc, op.ObjectName, bytes.NewReader(RandomData[:op.ObjectSize]), op.Bucket, int64(op.ObjectSize))
+		return putObject(HousekeepingSvc, op.ObjectName, bytes.NewReader(generateRandomBytes(op.ObjectSize)), op.Bucket, int64(op.ObjectSize))
 	}
 }
 
 // Prepare prepares the execution of the WriteOperation
-func (op WriteOperation) Prepare() error {
-	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).Trace("Preparing WriteOperation")
+func (op *WriteOperation) Prepare() error {
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Msg("Preparing WriteOperation")
 	return nil
 }
 
 // Prepare prepares the execution of the ListOperation
-func (op ListOperation) Prepare() error {
-	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).Trace("Preparing ListOperation")
-
-	if op.MPUEnabled {
-		if op.PartSize == 0 {
-			op.PartSize = uint64(s3manager.DefaultUploadPartSize)
-		}
-		if op.MPUConcurrency == 0 {
-			op.MPUConcurrency = s3manager.DefaultUploadConcurrency
-		}
-		return putObjectMPU(HousekeepingSvc, op.ObjectName, bytes.NewReader(RandomData[:op.ObjectSize]), op.Bucket, op.PartSize, op.MPUConcurrency)
-	} else {
-		return putObject(HousekeepingSvc, op.ObjectName, bytes.NewReader(RandomData[:op.ObjectSize]), op.Bucket, int64(op.ObjectSize))
-	}
-}
-
-// Prepare prepares the execution of the DeleteOperation
-func (op DeleteOperation) Prepare() error {
-	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).WithField("Preexisting?", op.WorksOnPreexistingObject).Trace("Preparing DeleteOperation")
+func (op *ListOperation) Prepare() error {
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Msg("Preparing ListOperation")
 	if op.WorksOnPreexistingObject {
 		return nil
 	}
 	if op.MPUEnabled {
 		if op.PartSize == 0 {
-			op.PartSize = uint64(s3manager.DefaultUploadPartSize)
+			op.PartSize = uint64(manager.DefaultUploadPartSize)
 		}
 		if op.MPUConcurrency == 0 {
-			op.MPUConcurrency = s3manager.DefaultUploadConcurrency
+			op.MPUConcurrency = manager.DefaultUploadConcurrency
 		}
-		return putObjectMPU(HousekeepingSvc, op.ObjectName, bytes.NewReader(RandomData[:op.ObjectSize]), op.Bucket, op.PartSize, op.MPUConcurrency)
+		uploader := manager.NewUploader(HousekeepingSvc, func(d *manager.Uploader) {
+			d.PartSize = int64(op.PartSize)
+			d.Concurrency = op.MPUConcurrency
+		})
+		return putObjectMPU(uploader, op.ObjectName, bytes.NewReader(generateRandomBytes(op.ObjectSize)), op.Bucket)
 	} else {
-		return putObject(HousekeepingSvc, op.ObjectName, bytes.NewReader(RandomData[:op.ObjectSize]), op.Bucket, int64(op.ObjectSize))
+		return putObject(HousekeepingSvc, op.ObjectName, bytes.NewReader(generateRandomBytes(op.ObjectSize)), op.Bucket, int64(op.ObjectSize))
+	}
+}
+
+// Prepare prepares the execution of the DeleteOperation
+func (op *DeleteOperation) Prepare() error {
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Bool("Preexisting?", op.WorksOnPreexistingObject).Msg("Preparing DeleteOperation")
+	if op.WorksOnPreexistingObject {
+		return nil
+	}
+	if op.MPUEnabled {
+		if op.PartSize == 0 {
+			op.PartSize = uint64(manager.DefaultUploadPartSize)
+		}
+		if op.MPUConcurrency == 0 {
+			op.MPUConcurrency = manager.DefaultUploadConcurrency
+		}
+		uploader := manager.NewUploader(HousekeepingSvc, func(d *manager.Uploader) {
+			d.PartSize = int64(op.PartSize)
+			d.Concurrency = op.MPUConcurrency
+		})
+		return putObjectMPU(uploader, op.ObjectName, bytes.NewReader(generateRandomBytes(op.ObjectSize)), op.Bucket)
+	} else {
+		return putObject(HousekeepingSvc, op.ObjectName, bytes.NewReader(generateRandomBytes(op.ObjectSize)), op.Bucket, int64(op.ObjectSize))
 	}
 }
 
 // Prepare does nothing here
-func (op Stopper) Prepare() error {
+func (op *Stopper) Prepare() error {
 	return nil
 }
 
 // Do executes the actual work of the ReadOperation
-func (op ReadOperation) Do() error {
-	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).WithField("Preexisting?", op.WorksOnPreexistingObject).Trace("Doing ReadOperation")
+func (op *ReadOperation) Do() error {
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Bool("Preexisting?", op.WorksOnPreexistingObject).Msg("Doing ReadOperation")
 	var err error
 	var duration time.Duration
 	if op.MPUEnabled {
 		if op.PartSize == 0 {
-			op.PartSize = s3manager.DefaultDownloadPartSize
+			op.PartSize = manager.DefaultDownloadPartSize
 		}
 		if op.MPUConcurrency == 0 {
-			op.MPUConcurrency = s3manager.DefaultDownloadConcurrency
+			op.MPUConcurrency = manager.DefaultDownloadConcurrency
 		}
+		downloader := manager.NewDownloader(svc, func(d *manager.Downloader) {
+			d.PartSize = int64(op.PartSize)
+			d.Concurrency = op.MPUConcurrency
+		})
 		start := time.Now()
-		err = getObjectMPU(svc, op.ObjectName, op.Bucket, op.PartSize, op.MPUConcurrency)
+		err = getObjectMPU(*downloader, op.ObjectName, op.Bucket)
 		duration = time.Since(start)
 	} else {
 		start := time.Now()
@@ -205,7 +221,7 @@ func (op ReadOperation) Do() error {
 		duration = time.Since(start)
 	}
 	if err != nil {
-		log.WithError(err).Errorf("could not read object body for %s in bucket %s", op.ObjectName, op.Bucket)
+		log.Error().Err(err).Msgf("could not read object body for %s in bucket %s", op.ObjectName, op.Bucket)
 	}
 	promRTLatency.WithLabelValues(op.TestName, "GET").Observe(float64(duration.Milliseconds()))
 	if err != nil {
@@ -218,8 +234,8 @@ func (op ReadOperation) Do() error {
 }
 
 // Do executes the actual work of the WriteOperation
-func (op WriteOperation) Do() error {
-	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).Trace("Doing WriteOperation")
+func (op *WriteOperation) Do() error {
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Msg("Doing WriteOperation")
 	var err error
 	var duration time.Duration
 	if op.MPUEnabled {
@@ -229,48 +245,44 @@ func (op WriteOperation) Do() error {
 		if op.MPUConcurrency == 0 {
 			op.MPUConcurrency = s3manager.DefaultUploadConcurrency
 		}
+		uploader := manager.NewUploader(svc, func(d *manager.Uploader) {
+			d.PartSize = int64(op.PartSize)
+			d.Concurrency = op.MPUConcurrency
+		})
 		start := time.Now()
-		err = putObjectMPU(svc, op.ObjectName, bytes.NewReader(RandomData[:op.ObjectSize]), op.Bucket, op.PartSize, op.MPUConcurrency)
+		err = putObjectMPU(uploader, op.ObjectName, bytes.NewReader(generateRandomBytes(op.ObjectSize)), op.Bucket)
 		duration = time.Since(start)
 	} else {
 		start := time.Now()
-		err = putObject(svc, op.ObjectName, bytes.NewReader(RandomData[:op.ObjectSize]), op.Bucket, int64(op.ObjectSize))
+		err = putObject(svc, op.ObjectName, bytes.NewReader(generateRandomBytes(op.ObjectSize)), op.Bucket, int64(op.ObjectSize))
 		duration = time.Since(start)
 	}
 	if err != nil {
-		if aerr, ok := err.(awserr.Error); ok && aerr.Code() == request.CanceledErrorCode {
-			// If the SDK can determine the request or retry delay was canceled
-			// by a context the CanceledErrorCode error code will be returned.
-			log.WithError(aerr).Errorf("Upload canceled due to timeout")
-		} else {
-			log.WithError(err).WithField("object", op.ObjectName).WithField("bucket", op.Bucket).Errorf("Failed to upload object,")
-		}
+		log.Error().Err(err).Str("object", op.ObjectName).Str("bucket", op.Bucket).Msg("Failed to upload object,")
+		promFailedOps.WithLabelValues(op.TestName, "PUT").Inc()
 		return err
 	} else {
-		log.WithField("bucket", op.Bucket).WithField("key", op.ObjectName).Tracef("Upload successful")
+		log.Trace().Str("bucket", op.Bucket).Str("key", op.ObjectName).Msg("Upload successful")
 	}
 	promRTLatency.WithLabelValues(op.TestName, "PUT").Observe(float64(duration.Milliseconds()))
-	if err != nil {
-		promFailedOps.WithLabelValues(op.TestName, "PUT").Inc()
-	} else {
-		promFinishedOps.WithLabelValues(op.TestName, "PUT").Inc()
-	}
+	promFinishedOps.WithLabelValues(op.TestName, "PUT").Inc()
 	promUploadedBytes.WithLabelValues(op.TestName, "PUT").Add(float64(op.ObjectSize))
 	return err
 }
 
 // Do executes the actual work of the ListOperation
-func (op ListOperation) Do() error {
-	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).Trace("Doing ListOperation")
+func (op *ListOperation) Do() error {
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Msg("Doing ListOperation")
 	start := time.Now()
-	_, err := ListObjects(svc, op.ObjectName, op.Bucket)
+	var err error
+	if op.UseV2 {
+		_, err = ListObjectsV2(svc, op.ObjectName, op.Bucket)
+	} else {
+		_, err = ListObjects(svc, op.ObjectName, op.Bucket)
+	}
 	duration := time.Since(start)
 	if err != nil {
-		// Cast err to awserr.Error to handle specific error codes.
-		aerr, ok := err.(awserr.Error)
-		if ok && aerr.Code() == s3.ErrCodeNoSuchKey {
-			log.WithError(aerr).Errorf("Could not find prefix %s in bucket %s when querying properties", op.ObjectName, op.Bucket)
-		}
+		log.Error().Err(err).Msgf("Could not find prefix %s in bucket %s when querying properties", op.ObjectName, op.Bucket)
 	}
 	promRTLatency.WithLabelValues(op.TestName, "LIST").Observe(float64(duration.Milliseconds()))
 	if err != nil {
@@ -282,8 +294,8 @@ func (op ListOperation) Do() error {
 }
 
 // Do executes the actual work of the DeleteOperation
-func (op DeleteOperation) Do() error {
-	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).Trace("Doing DeleteOperation")
+func (op *DeleteOperation) Do() error {
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Msg("Doing DeleteOperation")
 	start := time.Now()
 	err := deleteObject(svc, op.ObjectName, op.Bucket)
 	duration := time.Since(start)
@@ -291,7 +303,7 @@ func (op DeleteOperation) Do() error {
 		// Cast err to awserr.Error to handle specific error codes.
 		aerr, ok := err.(awserr.Error)
 		if ok && aerr.Code() == s3.ErrCodeNoSuchKey {
-			log.WithError(aerr).Errorf("Could not find object %s in bucket %s for deletion", op.ObjectName, op.Bucket)
+			log.Error().Err(aerr).Msgf("Could not find object %s in bucket %s for deletion", op.ObjectName, op.Bucket)
 		}
 	}
 	promRTLatency.WithLabelValues(op.TestName, "DELETE").Observe(float64(duration.Milliseconds()))
@@ -304,70 +316,73 @@ func (op DeleteOperation) Do() error {
 }
 
 // Do does nothing here
-func (op Stopper) Do() error {
+func (op *Stopper) Do() error {
 	return nil
 }
 
 // Clean removes the objects and buckets left from the previous ReadOperation
-func (op ReadOperation) Clean() error {
+func (op *ReadOperation) Clean() error {
 	if op.WorksOnPreexistingObject {
 		return nil
 	}
-	log.WithField("bucket", op.Bucket).WithField("object", op.ObjectName).WithField("Preexisting?", op.WorksOnPreexistingObject).Trace("Cleaning up ReadOperation")
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Bool("Preexisting?", op.WorksOnPreexistingObject).Msg("Cleaning up ReadOperation")
 	return deleteObject(HousekeepingSvc, op.ObjectName, op.Bucket)
 }
 
 // Clean removes the objects and buckets left from the previous WriteOperation
-func (op WriteOperation) Clean() error {
+func (op *WriteOperation) Clean() error {
 	return deleteObject(HousekeepingSvc, op.ObjectName, op.Bucket)
 }
 
 // Clean removes the objects and buckets left from the previous ListOperation
-func (op ListOperation) Clean() error {
+func (op *ListOperation) Clean() error {
+	if op.WorksOnPreexistingObject {
+		return nil
+	}
+	log.Trace().Str("bucket", op.Bucket).Str("object", op.ObjectName).Bool("Preexisting?", op.WorksOnPreexistingObject).Msg("Cleaning up ListOperation")
 	return deleteObject(HousekeepingSvc, op.ObjectName, op.Bucket)
 }
 
 // Clean removes the objects and buckets left from the previous DeleteOperation
-func (op DeleteOperation) Clean() error {
+func (op *DeleteOperation) Clean() error {
 	return nil
 }
 
 // Clean does nothing here
-func (op Stopper) Clean() error {
+func (op *Stopper) Clean() error {
 	return nil
 }
 
 // DoWork processes the workitems in the workChannel until
 // either the time runs out or a stopper is found
-func DoWork(workChannel chan WorkItem, doneChannel chan bool) {
+func DoWork(workChannel chan WorkItem, notifyChan <-chan struct{}, wg *sync.WaitGroup) {
+	defer wg.Done()
 	for {
 		select {
-		case <-WorkContext.Done():
-			log.Debugf("Runtime over - Got timeout from work context")
-			doneChannel <- true
+		case <-notifyChan:
+			log.Debug().Msg("Runtime over - Got timeout from work context")
 			return
 		case work := <-workChannel:
 			switch work.(type) {
-			case Stopper:
-				log.Debug("Found the end of the work Queue - stopping")
-				doneChannel <- true
+			case *Stopper:
+				log.Debug().Msg("Found the end of the work Queue - stopping")
 				return
 			}
 			err := work.Do()
 			if err != nil {
-				log.WithError(err).Error("Issues when performing work - ignoring")
+				log.Error().Err(err).Msg("Issues when performing work - ignoring")
 			}
 		}
 	}
 }
 
-func GenerateRandomBytes(size uint64) []byte {
+func generateRandomBytes(size uint64) []byte {
 	now := time.Now()
 	random := make([]byte, size)
 	n, err := rand.Read(random)
 	if err != nil {
-		log.WithError(err).Fatal("I had issues getting my random bytes initialized")
+		log.Fatal().Err(err).Msg("error generating random bytes")
 	}
-	log.Tracef("Generated %d random bytes in %v", n, time.Since(now))
+	log.Trace().Msgf("Generated %d random bytes in %v", n, time.Since(now))
 	return random
 }
